Split up UDP and TCP synch socket handling to a) make reusable and b) make code easier to follow.
